Update postinstaller.sh
Changed the order of operations at the end of the script to unmount and delete temporary files BEFORE checking to see if we need to kill the login window.

Added check for user and if the Setup Assistant is still running code.  Thanks to Richard Purves for that.  Send him tea.